data_extraction-main/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
google-cloud-translate>=3.15.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0
pydantic>=2.5.0